aider: Add a Gemfile.
# Aider chat conversation:

USER: add a Gemfile

ASSISTANT: Sure, we can add a Gemfile. A Gemfile is used in Ruby projects to manage gem dependencies.

Here's a basic Gemfile that specifies the use of Ruby 2.7.0 and includes the 'rake' gem, which is commonly used in Ruby projects. If you need other gems, you can add them to this file.

```ruby
new_file/Gemfile
<<<<<<< ORIGINAL

=======
source 'https://rubygems.org'
ruby '2.7.0'

gem 'rake'
>>>>>>> UPDATED
```

This Gemfile specifies that we're using Ruby 2.7.0 and that we're getting our gems from RubyGems.org. It also includes the 'rake' gem as a dependency. If you need other gems, you can add them to this file.