aider: Save embeddings in JSON files in the embeddings directory, including the SHA of the description.
# Aider chat conversation:

USER: instead of printing the embedding, save them in json files in the embedding directory. the json files should contain the sha of the description as well.

ASSISTANT: To save the embeddings in JSON files in the `embeddings` directory, we need to modify the `generate_embeddings.rb` script. We will replace the `puts` statement with code that writes the response to a JSON file. We will also add the SHA of the description to the JSON file.

Here are the changes:

```ruby
generate_embeddings.rb
<<<<<<< ORIGINAL
  # Print the response
  puts response.body
=======
  # Save the response to a JSON file
  sha = Digest::SHA256.hexdigest(description)
  File.open("embeddings/#{sha}.json", 'w') do |f|
    f.write(JSON.pretty_generate(JSON.parse(response.body).merge({"sha" => sha})))
  end
>>>>>>> UPDATED
```

This change will write the response from the OpenAI API, along with the SHA of the description, to a JSON file in the `embeddings` directory. The name of the JSON file will be the SHA of the description.